Moving oauth method to ww lib
