+++ lib/ww.js
@@ -1,4 +1,5 @@
 const request = require("request");
+const crypto = require("crypto");
 
 module.exports.sendMessage = function (msg, color, url, space, token) {
   const appMessage = {
@@ -57,6 +58,66 @@ module.exports.makeGraphQLCall = function (body, token, url, callback) {
     } else {
       console.log("GraphQL request succeeded with " + body);
       callback(null, JSON.parse(body));
+    }
+  });
+}
+
+///////////////////////////////////////////////////////////////////////////////
+//                                                                           //
+// Verification function                                                     //
+//                                                                           //
+///////////////////////////////////////////////////////////////////////////////
+module.exports.verifyWorkspace = function (response, challenge, secret) {
+
+  // Create the object that is going to be sent back to Workspace for
+  // verification.
+  var bodyChallenge = {
+    // This is what we end up hashing
+    "response": challenge
+  };
+
+  var responseBodyString = JSON.stringify(bodyChallenge);
+
+  var tokenForVerification = crypto
+    // Use the webhook secret as hash key
+    .createHmac("sha256", secret)
+    // and hash the body ("response": challenge)
+    .update(responseBodyString)
+    // ending up with the hex formatted hash.
+    .digest("hex");
+  // Write our response headers
+  response.writeHead(200, {
+    "Content-Type": "application/json; charset=utf-8",
+    "X-OUTBOUND-TOKEN": tokenForVerification
+  });
+
+  // Add our body and send it off.
+  response.end(responseBodyString);
+  console.log("Webhook was verified");
+}
+
+///////////////////////////////////////////////////////////////////////////////
+//                                                                           //
+// Obtain a token for oAuth                                                  //
+//                                                                           //
+///////////////////////////////////////////////////////////////////////////////
+module.exports.getToken = function (url, id, secret) {
+  request({
+      url: WWS_URL + '/oauth/token',
+      method: 'POST',
+      auth: {
+        user: APP_ID,
+        pass: APP_SECRET
+      },
+      form: {
+        'grant_type': 'client_credentials'
+      }
+  }, function(err, res) {
+    if (err) {
+      console.log("Failed to obtain Oauth token", err);
+    } else {
+      console.log("Here's the result from the oauth call");
+      console.log(res);
     }
   });
 }
